requirements/base.txt
# Base
# ------------------------------------------------------------------------------
psycopg2-binary==2.9.3
asyncpg==0.25.0
tortoise-orm==0.18.1
aerich==0.6.2

passlib==1.7.4
bcrypt==3.2.0

uvicorn[standard]==0.17.1

# FastAPI
# ------------------------------------------------------------------------------
fastapi==0.73.0
pydantic[email]==1.9.0
fastapi-mail==1.0.4